+++ components/refund.py
@@ -3,8 +3,15 @@ import numpy as np
 # 回收相关的模块
 
 
+def overwrite(key):
+    print(f'Overwriting {key}')
+    if key == 'default_oks':
+        st.session_state.overwrite_default_ok = True
+    else:
+        raise ValueError(f'Invalid key: {key}')
 
-def enemies(enemies,hits_dist:list[int],container=None):
+
+def enemies(enemies,hits_dist:list[int],container=None,default_oks:int=1):
     '''
     敌人相关
     
@@ -20,8 +27,8 @@ def enemies(enemies,hits_dist:list[int],container=None):
         result = []
         if not container:
             for idx,panel in enumerate(enemy_panels):
-                mod = panel.number_input(f'敌人{idx+1}回收补正', value=1.0, min_value=0.1, max_value=1.5,)
-                overkills = panel.slider(f'敌方{idx+1}过量hit数', min_value=0,max_value=max_overkills,step=1,value=1)
+                mod = panel.number_input(f'敌人{idx+1}回收补正', value=1.0, min_value=0.1, max_value=1.5,)                
+                overkills = panel.slider(f'敌方{idx+1}过量hit数', min_value=0,max_value=max_overkills,step=1,value=1)                                        
                 result.append((enemies[idx], mod, overkills))
     
     return result
