+++ pages/home.py
@@ -1,0 +1,3 @@
+import streamlit as st
+
+st.title('主页')
