+++ pages/shuffle.py
@@ -1,0 +1,4 @@
+import streamlit as st
+
+
+st.title('发牌概率计算器')
